some fixes

+++ src/main/java/ru/bengo/animaltracking/service/impl/AnimalTypeServiceImpl.java
@@ -6,9 +6,9 @@ import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
+import ru.bengo.animaltracking.dto.AnimalTypeDto;
 import ru.bengo.animaltracking.exception.AnimalTypeAlreadyExist;
 import ru.bengo.animaltracking.exception.AnimalTypeNotFoundException;
 import ru.bengo.animaltracking.model.AnimalType;
@@ -27,11 +27,12 @@ public class AnimalTypeServiceImpl implements AnimalTypeService {
     private static final Logger log = LoggerFactory.getLogger(AnimalTypeServiceImpl.class);
 
     @Override
-    public AnimalType create(@Valid AnimalType animalType) throws AnimalTypeAlreadyExist {
-        if (isAnimalTypeWithTypeExist(animalType.getType())) {
+    public AnimalType create(@Valid AnimalTypeDto animalTypeDto) throws AnimalTypeAlreadyExist {
+        if (isAnimalTypeWithTypeExist(animalTypeDto.type())) {
             throw new AnimalTypeAlreadyExist(Message.ANIMAL_TYPE_EXIST.getInfo());
         }
-        return animalTypeRepository.save(animalType);
+
+        return animalTypeRepository.save(convertToEntity(animalTypeDto));
     }
 
     @Override
@@ -46,26 +47,39 @@ public class AnimalTypeServiceImpl implements AnimalTypeService {
     }
 
     @Override
-    public AnimalType update(Long id, AnimalType animalType) throws AnimalTypeAlreadyExist, AnimalTypeNotFoundException {
+    public AnimalType update(Long id, AnimalTypeDto animalTypeDto) throws AnimalTypeAlreadyExist, AnimalTypeNotFoundException {
         Optional<AnimalType> foundAnimalType = animalTypeRepository.findById(id);
 
         if (foundAnimalType.isEmpty()) {
             throw new AnimalTypeNotFoundException(Message.ANIMAL_TYPE_NOT_FOUND.getInfo());
         }
-        if (isAnimalTypeWithTypeExist(animalType.getType())) {
+        if (isAnimalTypeWithTypeExist(animalTypeDto.type())) {
             throw new AnimalTypeAlreadyExist(Message.ANIMAL_TYPE_EXIST.getInfo());
         }
 
+        AnimalType animalType = convertToEntity(animalTypeDto);
         animalType.setId(id);
         return animalTypeRepository.save(animalType);
     }
 
     @Override
-    public void delete(@NotNull @Positive Long id) {
+    public void delete(@NotNull @Positive Long id) throws AnimalTypeNotFoundException {
+        Optional<AnimalType> foundAnimalType = animalTypeRepository.findById(id);
+
+        if (foundAnimalType.isEmpty()) {
+            throw new AnimalTypeNotFoundException(Message.ANIMAL_TYPE_NOT_FOUND.getInfo());
+        }
+
         animalTypeRepository.deleteById(id);
     }
 
     private boolean isAnimalTypeWithTypeExist(String type) {
         return animalTypeRepository.findByType(type).isPresent();
+    }
+
+    private AnimalType convertToEntity(AnimalTypeDto animalTypeDto) {
+        return AnimalType.builder()
+                .type(animalTypeDto.type())
+                .build();
     }
 }
